perf: stream profile page sections independently for progressive cold loads
Split app/[username]/page.tsx into three independently-Suspense-wrapped async
sections so a cold load streams each as its own data resolves instead of
blocking on the full profile fill:

- Identity header — getGitHubUserData + curation row (~300-500ms), owns the
  404 and hard-error decisions so they land first
- Contribution graph — getContributionCalendar
- Stats line + PR rail — the slow getProfileData path (search, repo
  breakdown, featured walk)

The page body awaits nothing; each section resolves params and its data inside
its own boundary, so the skeleton fallbacks (mirroring loading.tsx) form the
prerendered static shell. Warm loads are unchanged. connection() still guards
every error/degraded branch so a transient failure never becomes the cached
shell.

Fix getGitHubUserData return type (GithubUser, not the lean GitHubUser) — the
/users endpoint returns the full profile; the old type only compiled because
getProfileData re-cast it and no caller read it directly.
